multi-deed

+++ 10. Multi-Deed/contracts/Multi-Deed.sol
@@ -4,16 +4,16 @@ pragma solidity ^0.8.0;
 contract DeedMultiPayout {
     address public lawyer;
     address payable public beneficiary;
-    uint public earliest;
-    uint public amount;
-    uint public constant PAYOUTS = 10;
-    uint public constant INTERVAL = 10;
-    uint public paidPayouts;
+    uint256 public earliest;
+    uint256 public amount;
+    uint256 public constant PAYOUTS = 10;
+    uint256 public constant INTERVAL = 10;
+    uint256 public paidPayouts;
 
     constructor(
         address _lawyer,
         address payable _beneficiary,
-        uint fromNow
+        uint256 fromNow
     ) payable {
         lawyer = _lawyer;
         beneficiary = _beneficiary;
@@ -26,8 +26,8 @@ contract DeedMultiPayout {
         require(block.timestamp >= earliest, "too early");
         require(paidPayouts < PAYOUTS, "no payout left");
 
-        uint elligiblePayouts = (block.timestamp - earliest) / INTERVAL;
-        uint duePayouts = elligiblePayouts - paidPayouts;
+        uint256 elligiblePayouts = (block.timestamp - earliest) / INTERVAL;
+        uint256 duePayouts = elligiblePayouts - paidPayouts;
         duePayouts = duePayouts + paidPayouts > PAYOUTS
             ? PAYOUTS - paidPayouts
             : duePayouts;
